+++ src/Enigma.GitClient.App/DependencyInjection/ServiceCollectionExtensions.cs
@@ -87,6 +87,7 @@ public static class ServiceCollectionExtensions
             services.AddTransient<TagsPageView>();
             services.AddTransient<RemotesPageView>();
             services.AddTransient<ConflictResolutionPageView>();
+            services.AddTransient<IdentityPageView>();
             services.AddTransient<IntegrationsPageView>();
             services.AddTransient<SettingsPageView>();
 
@@ -97,6 +98,7 @@ public static class ServiceCollectionExtensions
             services.AddSingleton<TagsPageViewModel>();
             services.AddSingleton<RemotesPageViewModel>();
             services.AddSingleton<ConflictResolutionPageViewModel>();
+            services.AddSingleton<IdentityPageViewModel>();
             services.AddSingleton<IntegrationsPageViewModel>();
             services.AddSingleton<SettingsPageViewModel>();